Move contents of index.h into merge.h and remove the former
